Introduce Drone plugin
This contains the beginnins of a Drone plugin that launches the XMPP Interop Framework tests against an XMPP server.

The new plugin takes the Docker image that is produced by the XMPP Interop Framework, extends that to be a Drone Plugin image, and invokes the tests within it using a small script.

All of this is almost an exact copy of the work that was done for Woodpecker by Dan.